moves convert_html method to coredocument
